Update CsWin32 and migrate the rest of ComCtl32
