--- src/System.Windows.Forms.Primitives/src/Interop/ComCtl32/Interop.TVINSERTSTRUCTW.cs
@@ -1,15 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-internal static partial class Interop
-{
-    internal static partial class ComCtl32
-    {
-        public struct TVINSERTSTRUCTW
-        {
-            public IntPtr hParent;
-            public IntPtr hInsertAfter;
-            public TVITEMW item;
-        }
-    }
-}
