+++ src/System.Windows.Forms.Primitives/src/NativeMethods.txt
@@ -517,7 +517,7 @@ LVGGR_*
 LVGROUP
 LVFINDINFOW
 LVFINDINFOW_FLAGS
-LVHITTESTINFO_FLAGS
+LVHITTESTINFO
 LVINSERTMARK
 LVIR_*
 LVM_*
@@ -791,9 +791,11 @@ TREE_VIEW_ITEM_STATE_FLAGS
 TTDT_*
 TTM_*
 TTN_*
+TTTOOLINFOW
 TTS_*
 TVGN_*
 TVHITTESTINFO
+TVINSERTSTRUCTW
 TVITEM_MASK
 TVITEMW
 TVI_*
