+++ src/System.Windows.Forms/src/System/Windows/Forms/ToolTip/ToolTip.cs
@@ -4,7 +4,6 @@
 using System.ComponentModel;
 using System.Drawing;
 using static Interop;
-using static Interop.ComCtl32;
 
 namespace System.Windows.Forms;
